Conflicts are resolved

+++ tipos_producto/service.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm import Session
 from schemas import Respuesta
 import tipos_producto.models as models
 import tipos_producto.schemas as schemas
+import productos.models as producto_models
 
 def create_tipo_producto(db: Session, tipo_producto: schemas.Tipo_ProductoCrear):
     db_tipo_producto = models.Tipo_Producto(nombre=tipo_producto.nombre, descripcion=tipo_producto.descripcion, funcionalidad=tipo_producto.funcionalidad)
@@ -26,6 +27,13 @@ def get_tipos_producto(db: Session):
     returned = db.query(models.Tipo_Producto).all()
 
     respuesta = Respuesta[list[schemas.Tipo_Producto]](ok=True, mensaje='Tipos de producto encontrados', data=returned)
+    tipos = []
+
+    for tipo in returned:
+        tip = schemas.Tipo_Producto(nombre=tipo.nombre, descripcion=tipo.descripcion,  funcionalidad=tipo.funcionalidad, id=tipo.id)
+        tipos.append(tip)
+
+    respuesta = Respuesta[list[schemas.Tipo_Producto]](ok=True, mensaje='Tipos de producto encontrados', data=tipos)
     return respuesta
 
 
@@ -52,6 +60,10 @@ def delete_tipo_producto(db: Session, id: int):
     if tipo_productoFound == None:
         return Respuesta[schemas.Tipo_Producto](ok=False, mensaje='Tipo de producto a eliminar no encontrado')
 
+    productos = db.query(producto_models.Producto).filter(producto_models.Producto.tipo_producto_id == id).first()
+
+    if productos:
+        return Respuesta[schemas.Categoria](ok=False, mensaje='No se puede eliminar una categoría que esté siendo usada por un producto.')
 
     db.query(models.Tipo_Producto).filter(models.Tipo_Producto.id == id).delete()
     db.commit()
